feat: update testing structure
